+++ src/gd_cubism_effect_custom.hpp
@@ -52,19 +52,19 @@ public:
         this->_initialized = false;
     }
 
-    virtual void _cubism_prologue(InternalCubismUserModel* model, const float delta) override {
+    virtual void _cubism_prologue(InternalCubismUserModel* model, const double delta) override {
         if(this->_initialized == false) return;
         if(this->_active == false) return;
         this->emit_signal("cubism_prologue", model->_owner_viewport, delta);
     }
 
-    virtual void _cubism_process(InternalCubismUserModel* model, const float delta) override {
+    virtual void _cubism_process(InternalCubismUserModel* model, const double delta) override {
         if(this->_initialized == false) return;
         if(this->_active == false) return;
         this->emit_signal("cubism_process", model->_owner_viewport, delta);
     }
 
-    virtual void _cubism_epilogue(InternalCubismUserModel* model, const float delta) override {
+    virtual void _cubism_epilogue(InternalCubismUserModel* model, const double delta) override {
         if(this->_initialized == false) return;
         if(this->_active == false) return;
         this->emit_signal("cubism_epilogue", model->_owner_viewport, delta);
