+++ src/gd_cubism_effect_target_point.hpp
@@ -170,7 +170,7 @@ public:
         this->_initialized = false;
     }
 
-    virtual void _cubism_process(InternalCubismUserModel* model, const float delta) override {
+    virtual void _cubism_process(InternalCubismUserModel* model, const double delta) override {
         if(this->_initialized == false) return;
         if(this->_active == false) return;
         if(this->_target_point == nullptr) return;
